requirements.txt
PyQt5
PyQt5-Qt5
PyQt5-sip
PyQt5-stubs
python-dotenv
pyyaml
black